Separate app files from supporting files
