--- .npmrc
@@ -1,2 +1,0 @@
-# Expose Astro dependencies for `pnpm` users
-shamefully-hoist=true
