+++ app/.npmrc
@@ -1,0 +1,2 @@
+# Expose Astro dependencies for `pnpm` users
+shamefully-hoist=true
