+++ app/public/old/index-old.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Lex von Biedenfeld</title>
+  <link rel="stylesheet" href="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-core/11.0.0/css/fabric.min.css" />
+  <link rel="stylesheet" href="styles/reset.css">
+  <link rel="stylesheet" href="styles/main.css">
+</head>
+<body>
+<header>
+
+</header>
+<main role="main">
+  <header id="banner">
+    <div class="text_content">
+      <h1><span>Welcome</span></h1>
+      <h2><span>What can be today?</span><br/><span>What will be tomorrow?</span></h2>
+    </div>
+    <img id="avatar" src="#" alt="A photo of me" sizes="" srcset="" />
+  </header>
+  <nav class="image_nav">
+    <a class="image_button" href="#" title="About Me"><img src="images/test-background2.jpg" alt="About Me"/>
+    </a><a class="image_button" href="#" title="My Projects"><img src="images/test-background2.jpg" alt="My Projects"/>
+    </a><a class="image_button" href="#" title="Other"><img src="images/test-background2.jpg" alt="Other"/>
+    </a><a class="image_button" href="#" title="Contact Me"><img src="images/test-background2.jpg" alt="Contact Me"/></a>
+  </nav>
+</main>
+<footer id="site_foot">
+  <small>&copy; 2020 Lex von Biedenfeld.</small>
+  <small>Last updated <time datetime="2020-03-06">Mar 6, 2019</time>.</small>
+</footer>
+<script src="scripts/app.js"></script>
+</body>
+</html>
