+++ app/src/components/BioCard.astro
