+++ app/src/components/Footer.astro
