+++ app/src/components/head.html
@@ -1,0 +1,7 @@
+<head>
+  <meta charset="UTF-8">
+  <title>{% if include.title %}{{ include.title }} | {% endif %}Lex</title>
+  <link rel="stylesheet" href="https://static2.sharepointonline.com/files/fabric/office-ui-fabric-core/11.0.0/css/fabric.min.css" />
+  <link rel="stylesheet" href="styles/reset.css" />
+  <link rel="stylesheet" href="styles/main.css" />
+</head>
