+++ app/src/layouts/BaseLayout.astro
@@ -1,0 +1,35 @@
+---
+import { SITE_TITLE, SITE_DESCRIPTION } from '@utils/config';
+import '@assets/reset.css';
+import '@assets/main.css';
+
+export interface Props {
+	title?: string;
+	description?: string;
+	image?: string;
+}
+
+const { title, description = SITE_DESCRIPTION } = Astro.props;
+---
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta http-equiv="X-UA-Compatible" content="IE=edge">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+	<!-- Primary Meta Tags -->
+	<title>{SITE_TITLE}{title ? ` | ${title}`: ''}</title>
+	<meta name="title" content={title} />
+	<meta name="description" content={description} />
+	<!-- Open Graph / Facebook -->
+	<meta property="og:type" content="website" />
+	<meta property="og:url" content={Astro.url} />
+	<meta property="og:title" content={title} />
+	<meta property="og:description" content={description} />
+</head>
+<body>
+	<slot/>
+</body>
+</html>
