--- src/utils/config.ts
@@ -1,5 +1,0 @@
-// Place any global data in this file.
-// You can import this data from anywhere in your site by using the `import` keyword.
-
-export const SITE_TITLE = 'Lex';
-export const SITE_DESCRIPTION = 'Welcome to my website!';
