--- tsconfig.json
@@ -1,13 +1,0 @@
-{
-	"extends": "astro/tsconfigs/base",
-	"compilerOptions": {
-		"baseUrl": ".",
-		"paths": {
-			"@components/*": ["src/components/*"],
-			"@layouts/*": ["src/layouts/*"],
-			"@assets/*": ["src/assets/*"],
-			"@utils/*": ["src/utils/*"]
-		},
-		"types": ["astro/client"]
-	}
-}
